third_party/blink/web_tests/fast/xpath/4XPath/Borrowed/od_20000608-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
PASS nodeset.snapshotLength is 0
PASS nodeset.snapshotLength is 12
PASS successfullyParsed is true

TEST COMPLETE
